+++ CHANGELOG.md
@@ -15,6 +15,8 @@ PyCheval adheres to [semantic versioning](https://semver.org/).
 
 ### Fixed
 
+- Trade parties can a maximum of one trade contact. Replace
+  `TradeParty.contacts` with `TradeParty.contact`.
 - Validate that the seller has a tax registration number.
 
 ## [0.1.0] - 2025-05-07